功能优化与样式的完善

+++ src/components/TrashDetail.vue
@@ -53,6 +53,10 @@ export default {
     this.getTrashNotes()
       .then(()=>{
         this.setCurTrashNote({curTrashNotedId: this.$route.query.noteId})
+        this.$router.replace({
+          path:'/trash',
+          query:{noteId: this.curTrashNote.id}
+        })
       })
   },
 
@@ -81,10 +85,29 @@ export default {
     ]),
 
     onDelete(){
-      this.deleteTrashNote({noteId: this.curTrashNote.id})
+      this.$confirm('彻底删除后无法恢复', '确定删除吗？', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(()=>{
+        return this.deleteTrashNote({noteId: this.curTrashNote.id})
+      }).then(()=>{
+        this.setCurTrashNote()
+        this.$router.replace({
+          path:'/trash',
+          query:{noteId: this.curTrashNote.id}
+        })
+      })
     },
     onRevert(){
       this.revertTrashNote({noteId: this.curTrashNote.id})
+        .then(()=>{
+          this.setCurTrashNote()
+          this.$router.replace({
+            path:'/trash',
+            query:{noteId: this.curTrashNote.id}
+          })
+        })
     },
   },
   beforeRouteUpdate(to, from, next){
